Move repo-specific functions to repos module.
Also fix worker update dealing with existing source repo checkouts.
