+++ lib/osv/__init__.py
@@ -15,5 +15,6 @@
 
 from .bug import *
 from .impact import *
+from .repos import *
 from .sources import *
 from .types import *
